refactor(ingestion): consolidate ProfilerRequest and GE error message into profiling.common
