+++ metadata-ingestion/src/datahub/ingestion/source/sqlalchemy_profiler/sqlalchemy_profiler.py
@@ -214,7 +214,7 @@ def _is_single_row_query_method(query: Any) -> bool:
 
 
 if TYPE_CHECKING:
-    from datahub.ingestion.source.profiling.profiler_request import ProfilerRequest
+    from datahub.ingestion.source.profiling.common import ProfilerRequest
 
 logger: logging.Logger = logging.getLogger(__name__)
